Attempt to get a working prompt with git status notifications
The code works independent of each other part but stitching it
together is proving to be a bit of a problem. Right now jb_git_ps1
reports jb_git as being an empty array, though testing in the shell
shows that it has data. Perhaps jb_git_ps1 is being processed before
jb_git_chpwd_update has a chance to execute :(